adjusted e2e structure to allow for more flexibliity and better regression
