+++ src/applications/check-in/tests/e2e/production/errors/check-in-failed.cypress.spec.js
@@ -1,7 +1,7 @@
 import features from '../mocks/enabled.json';
-
-import mockCheckIn from '../../api/local-mock-api/mocks/check.in.response';
-import mockValidate from '../../api/local-mock-api/mocks/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/check.in.response';
+import mockValidate from '../../../../api/local-mock-api/mocks/validate.responses';
+import Timeouts from 'platform/testing/e2e/timeouts';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
@@ -9,7 +9,7 @@ describe('Check In Experience -- ', () => {
       req.reply(mockValidate.createMockSuccessResponse({}));
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(mockCheckIn.createMockSuccessResponse({}));
+      req.reply(mockCheckIn.createMockFailedResponse({}));
     });
     cy.intercept('GET', '/v0/feature_toggles*', features);
   });
@@ -18,20 +18,16 @@ describe('Check In Experience -- ', () => {
       window.sessionStorage.clear();
     });
   });
-  it('happy path', () => {
+  it('check in failed with a 200 and error message in the body', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
     cy.get('legend > h2').contains('information');
-    cy.injectAxe();
-    cy.axeCheck();
     cy.get('[data-testid="no-button"]').click();
     cy.get('h1').contains('Your appointment');
-    cy.injectAxe();
-    cy.axeCheck();
     cy.get('.usa-button').click();
-    cy.get('va-alert > h1').contains('checked in');
-    cy.injectAxe();
-    cy.axeCheck();
+    cy.get('h1', { timeout: Timeouts.slow })
+      .should('be.visible')
+      .and('contain', 'We couldn’t check you in');
   });
 });
