+++ src/applications/check-in/tests/e2e/production/errors/malformed.token.cypress.spec.js
@@ -1,16 +1,20 @@
 import features from '../mocks/enabled.json';
-
-import mockCheckIn from '../../api/local-mock-api/mocks/check.in.response';
-import mockValidate from '../../api/local-mock-api/mocks/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/check.in.response';
+import mockValidate from '../../../../api/local-mock-api/mocks/validate.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply(mockValidate.createMockSuccessResponse({}));
+      req.reply({
+        statusCode: 200,
+        body: mockValidate.createMockFailedResponse({}),
+        delay: 10, // milliseconds
+      });
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
       req.reply(mockCheckIn.createMockSuccessResponse({}));
     });
+
     cy.intercept('GET', '/v0/feature_toggles*', features);
   });
   afterEach(() => {
@@ -18,12 +22,9 @@ describe('Check In Experience -- ', () => {
       window.sessionStorage.clear();
     });
   });
-
-  it('should show error page since there is no data to load locally', () => {
-    const featureRoute = '/health-care/appointment-check-in/update-information';
-    cy.window().then(window => {
-      window.sessionStorage.clear();
-    });
+  it('token is not valid', () => {
+    const featureRoute =
+      '/health-care/appointment-check-in/?id=MALFORMED_TOKEN';
     cy.visit(featureRoute);
     cy.get('h1').contains('We couldn’t check you in');
   });
