+++ src/applications/check-in/tests/e2e/production/errors/server.404.on.check-in.cypress.spec.js
@@ -1,14 +1,14 @@
 import features from '../mocks/enabled.json';
-import mockCheckIn from '../../api/local-mock-api/mocks/check.in.response';
-import mockValidate from '../../api/local-mock-api/mocks/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/check.in.response';
+import mockValidate from '../../../../api/local-mock-api/mocks/validate.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply(500, mockValidate.createMockFailedResponse({}));
+      req.reply(mockValidate.createMockSuccessResponse({}));
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(mockCheckIn.createMockFailedResponse({}));
+      req.reply(404, mockCheckIn.createMockFailedResponse({}));
     });
     cy.intercept('GET', '/v0/feature_toggles*', features);
     cy.window().then(window => {
@@ -20,10 +20,14 @@ describe('Check In Experience -- ', () => {
       window.sessionStorage.clear();
     });
   });
-  it('validate - 500 api error', () => {
+  it('check in - 404 api error', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
+    cy.get('legend > h2').contains('information');
+    cy.get('[data-testid="no-button"]').click();
+    cy.get('h1').contains('Your appointment');
+    cy.get('.usa-button').click();
     cy.url().should('match', /error/);
     cy.get('h1').contains('We couldn’t check you in');
   });
