+++ src/applications/check-in/tests/e2e/production/session/no.session.to.reload.on.refresh.cypress.spec.js
@@ -1,31 +1,30 @@
 import features from '../mocks/enabled.json';
-import mockCheckIn from '../../api/local-mock-api/mocks/check.in.response';
-import mockValidate from '../../api/local-mock-api/mocks/validate.responses';
+
+import mockCheckIn from '../../../../api/local-mock-api/mocks/check.in.response';
+import mockValidate from '../../../../api/local-mock-api/mocks/validate.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply(403, mockValidate.createMockFailedResponse({}));
+      req.reply(mockValidate.createMockSuccessResponse({}));
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(200, mockCheckIn.createMockSuccessResponse({}));
+      req.reply(mockCheckIn.createMockSuccessResponse({}));
     });
     cy.intercept('GET', '/v0/feature_toggles*', features);
-    cy.window().then(window => {
-      window.sessionStorage.clear();
-    });
   });
   afterEach(() => {
     cy.window().then(window => {
       window.sessionStorage.clear();
     });
   });
-  it('check in - 404 api error', () => {
-    const featureRoute =
-      '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
-    cy.visit(featureRoute);
 
-    cy.url().should('match', /error/);
+  it('should show error page since there is no data to load locally', () => {
+    const featureRoute = '/health-care/appointment-check-in/update-information';
+    cy.window().then(window => {
+      window.sessionStorage.clear();
+    });
+    cy.visit(featureRoute);
     cy.get('h1').contains('We couldn’t check you in');
   });
 });
