+++ src/applications/check-in/tests/e2e/production/session/session.reloads.on.refresh.cypress.spec.js
@@ -1,7 +1,7 @@
 import features from '../mocks/enabled.json';
 
-import mockCheckIn from '../../api/local-mock-api/mocks/check.in.response';
-import mockValidate from '../../api/local-mock-api/mocks/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/check.in.response';
+import mockValidate from '../../../../api/local-mock-api/mocks/validate.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
@@ -12,31 +12,13 @@ describe('Check In Experience -- ', () => {
       req.reply(mockCheckIn.createMockSuccessResponse({}));
     });
     cy.intercept('GET', '/v0/feature_toggles*', features);
-    cy.window().then(window => {
-      const sample = JSON.stringify({
-        token: 'the-old-id',
-      });
-      window.sessionStorage.setItem(
-        'health.care.check-in.current.uuid',
-        sample,
-      );
-    });
   });
   afterEach(() => {
     cy.window().then(window => {
       window.sessionStorage.clear();
     });
   });
-  it('url is prioritized over session data', () => {
-    cy.window().then(window => {
-      const data = window.sessionStorage.getItem(
-        'health.care.check-in.current.uuid',
-      );
-      const sample = JSON.stringify({
-        token: 'the-old-id',
-      });
-      expect(data).to.equal(sample);
-    });
+  it('on page reload, the data should be pull from session storage and redirected to landing screen with data loaded', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
@@ -50,6 +32,11 @@ describe('Check In Experience -- ', () => {
         token: '46bebc0a-b99c-464f-a5c5-560bc9eae287',
       });
       expect(data).to.equal(sample);
+      cy.reload();
+      // redirected back to landing page to reload the data
+      cy.url().should('match', /id=46bebc0a-b99c-464f-a5c5-560bc9eae287/);
+
+      cy.get('legend > h2').contains('information');
     });
   });
 });
