+++ src/applications/check-in/tests/e2e/production/session/url.is.priotized.over.session.cypress.spec.js
@@ -1,7 +1,7 @@
 import features from '../mocks/enabled.json';
 
-import mockCheckIn from '../../api/local-mock-api/mocks/check.in.response';
-import mockValidate from '../../api/local-mock-api/mocks/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/check.in.response';
+import mockValidate from '../../../../api/local-mock-api/mocks/validate.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
@@ -12,13 +12,31 @@ describe('Check In Experience -- ', () => {
       req.reply(mockCheckIn.createMockSuccessResponse({}));
     });
     cy.intercept('GET', '/v0/feature_toggles*', features);
+    cy.window().then(window => {
+      const sample = JSON.stringify({
+        token: 'the-old-id',
+      });
+      window.sessionStorage.setItem(
+        'health.care.check-in.current.uuid',
+        sample,
+      );
+    });
   });
   afterEach(() => {
     cy.window().then(window => {
       window.sessionStorage.clear();
     });
   });
-  it('data is saved to session storage', () => {
+  it('url is prioritized over session data', () => {
+    cy.window().then(window => {
+      const data = window.sessionStorage.getItem(
+        'health.care.check-in.current.uuid',
+      );
+      const sample = JSON.stringify({
+        token: 'the-old-id',
+      });
+      expect(data).to.equal(sample);
+    });
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
