--- src/applications/check-in/todo.md
@@ -1,4 +1,0 @@
-# todo
-
-- update mocks
-- restructure test to be a v1 folder
